refactor: move away more source code generation code
