+++ jit-compiler/src/test/java/source/NameGenerator.java
@@ -1,6 +1,6 @@
-// Copyright (c) 2020-2024 Tesla (Yinsen) Zhang.
+// Copyright (c) 2020-2025 Tesla (Yinsen) Zhang.
 // Use of this source code is governed by the MIT license that can be found in the LICENSE.md file.
-package org.aya.compiler;
+package source;
 
 import org.aya.generic.Constants;
 import org.jetbrains.annotations.NotNull;
