+++ jit-compiler/src/test/java/source/SourceClassBuilder.java
@@ -3,7 +3,6 @@
 package source;
 
 import kala.collection.immutable.ImmutableSeq;
-import org.aya.compiler.SourceBuilder;
 import org.aya.compiler.free.*;
 import org.aya.compiler.free.data.FieldRef;
 import org.aya.compiler.free.data.MethodRef;
