+++ jit-compiler/src/test/java/source/SourceCodeBuilder.java
@@ -4,7 +4,6 @@ package source;
 
 import kala.collection.immutable.ImmutableSeq;
 import kala.collection.immutable.primitive.ImmutableIntSeq;
-import org.aya.compiler.SourceBuilder;
 import org.aya.compiler.free.ArgumentProvider;
 import org.aya.compiler.free.FreeCodeBuilder;
 import org.aya.compiler.free.FreeJavaExpr;
